requirements.txt
scipy~=1.10.1
pandas~=2.0.2
matplotlib~=3.7.1
numpy~=1.24.3
dtw-python~=1.3.0